:art: Rename utility 'templates' to 'template'
